requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
matplotlib>=3.4.3
pillow>=8.3.1
transformers>=4.11.3
einops>=0.3.2
tqdm>=4.62.3
scikit-learn>=1.0.0 